bash: Ensure we use glibc getcwd(), not builtin copy
My build system runs in a chroot, and then creates bind mounts
that point outside of the chroot.  The bash implementation of getcwd()
breaks badly on this.

glibc has a perfectly working version which on modern Linux simply
invokes the getcwd() system call.  However, when cross compiling,
bash's configure script gives up, and so defers to its built in
version.

I'm not sure whether dietlibc/uclibc will malloc as bash wants, so
this patch may not be applicable everywhere.  Probably a better fix
would be for bash to detect glibc's get_current_dir_name() and use it.