Add deprecated find.pl from Perl 5.14 to fix perlpath.pl
The perlpath.pl tool is being used from openssl.inc, and it requires
path.pl to be loadable. Unfortunately, versions of Perl newer than 5.14
do not longer include it. This makes it possible to build ostree-native
when the host system has a newer Perl.

Based on this patch for upstream Poky:
http://patches.openembedded.org/patch/29989/mbox/