util-linux: Update for swapoff being a separate binary
In older versions of util-linux, swapon and swapoff were the
same binary, and it did runtime detection.  But since v2.22
which is util-linux commit 6cf8d46ceefe9a7, they are separate
binaries.

This patch is necessary to make the util-linux version of
swapoff work at all - currently in OE swapoff = swapon, which
is clearly broken =)

Probably most OE consumers use the busybox swapoff and hence
this has gone unnoticed.

Signed-off-by: Colin Walters <walters@verbum.org>